(PC-11881)[API]:sendinblue booking_confirmation_email
